+++ noodles-vcf/src/header/record/value/map/info.rs
@@ -2,7 +2,7 @@ use std::fmt;
 
 use indexmap::IndexMap;
 
-use super::{tag, Described, Fields, Indexed, Inner, Map, TryFromFieldsError, Typed};
+use super::{builder, tag, Described, Fields, Indexed, Inner, Map, TryFromFieldsError, Typed};
 use crate::header::{
     info::{Key, Type},
     FileFormat, Number,
@@ -23,6 +23,7 @@ pub struct Info {
 impl Inner for Info {
     type Id = Key;
     type StandardTag = StandardTag;
+    type Builder = builder::TypedDescribedIndexed<Self>;
 }
 
 impl Typed for Info {
@@ -210,6 +211,27 @@ fn validate_type_fields(
     }
 
     Ok(())
+}
+
+impl builder::Inner<Info> for builder::TypedDescribedIndexed<Info> {
+    fn build(self) -> Result<Info, builder::BuildError> {
+        let number = self
+            .number
+            .ok_or(builder::BuildError::MissingField("Number"))?;
+
+        let ty = self.ty.ok_or(builder::BuildError::MissingField("Type"))?;
+
+        let description = self
+            .description
+            .ok_or(builder::BuildError::MissingField("Description"))?;
+
+        Ok(Info {
+            number,
+            ty,
+            description,
+            idx: self.idx,
+        })
+    }
 }
 
 #[cfg(test)]