+++ noodles-vcf/CHANGELOG.md
@@ -6,6 +6,8 @@
 
   * vcf/header/contig: Add name wrapper (`Name`).
 
+  * vcf/header/record/value/map: Add builder (`Builder<I>`).
+
   * vcf/header/record/value/map/contig: Add mutable getter for length
     (`Contig::length_mut`) ([#99]).